// src/app/shared/model/AppAssetCollateralForInsurance.Model.ts
export class AppAssetCollateralForInsuranceObj {
    AppInsObjId: number;
    AppAssetId: number;
    AppCollateralId: number;
    AssetSeqNo: number;
    InsSeqNo: number;
    CollateralNo: String;
    FullAssetName: String;
    InsuredBy: String;
    InsuredByCode: String;
    PaidBy: String;
    CoverPeriod: String;
    InscoName: String;
    InsLength: number;
    TotalCustPremiAmt: number;
    MainPremi: number;
    AddPremi: number;
    Fee: number;
    PaidAmtByCust: number;
    InsCpltzAmt: number;
    InsDiscAmt: number;
    Behaviour: String;
}